+++ D3DFramework/D3DFramework/KrugMini.cpp
@@ -10,7 +10,7 @@ KrugMini::KrugMini()
 
 	anim->AttachToDynamicMesh(dmesh);
 
-	anim->SetLoop(UnitState::IDLE1, true);
+	anim->SetLoop((int)UnitState::IDLE1, true);
 }
 
 KrugMini::~KrugMini()
